+++ src/Symfony/Component/Stopwatch/Section.php
@@ -29,6 +29,11 @@ class Section
     private $origin;
 
     /**
+     * @var bool
+     */
+    private $morePrecision;
+
+    /**
      * @var string
      */
     private $id;
@@ -41,11 +46,13 @@ class Section
     /**
      * Constructor.
      *
-     * @param float|null $origin Set the origin of the events in this section, use null to set their origin to their start time
+     * @param float|null $origin        Set the origin of the events in this section, use null to set their origin to their start time
+     * @param bool       $morePrecision If true, time is stored as float to keep the original microsecond precision
      */
-    public function __construct($origin = null)
+    public function __construct($origin = null, $morePrecision = false)
     {
         $this->origin = is_numeric($origin) ? $origin : null;
+        $this->morePrecision = $morePrecision;
     }
 
     /**
@@ -74,7 +81,7 @@ class Section
     public function open($id)
     {
         if (null === $session = $this->get($id)) {
-            $session = $this->children[] = new self(microtime(true) * 1000);
+            $session = $this->children[] = new self(microtime(true) * 1000, $this->morePrecision);
         }
 
         return $session;
@@ -113,7 +120,7 @@ class Section
     public function startEvent($name, $category)
     {
         if (!isset($this->events[$name])) {
-            $this->events[$name] = new StopwatchEvent($this->origin ?: microtime(true) * 1000, $category);
+            $this->events[$name] = new StopwatchEvent($this->origin ?: microtime(true) * 1000, $category, $this->morePrecision);
         }
 
         return $this->events[$name]->start();