+++ src/Symfony/Component/Translation/DependencyInjection/TranslationDumperPass.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
+namespace Symfony\Component\Translation\DependencyInjection;
 
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -20,15 +20,24 @@ use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
  */
 class TranslationDumperPass implements CompilerPassInterface
 {
+    private $writerServiceId;
+    private $dumperTag;
+
+    public function __construct($writerServiceId = 'translation.writer', $dumperTag = 'translation.dumper')
+    {
+        $this->writerServiceId = $writerServiceId;
+        $this->dumperTag = $dumperTag;
+    }
+
     public function process(ContainerBuilder $container)
     {
-        if (!$container->hasDefinition('translation.writer')) {
+        if (!$container->hasDefinition($this->writerServiceId)) {
             return;
         }
 
-        $definition = $container->getDefinition('translation.writer');
+        $definition = $container->getDefinition($this->writerServiceId);
 
-        foreach ($container->findTaggedServiceIds('translation.dumper', true) as $id => $attributes) {
+        foreach ($container->findTaggedServiceIds($this->dumperTag, true) as $id => $attributes) {
             $definition->addMethodCall('addDumper', array($attributes[0]['alias'], new Reference($id)));
         }
     }