+++ src/Symfony/Component/Translation/DependencyInjection/TranslatorPass.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
+namespace Symfony\Component\Translation\DependencyInjection;
 
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -18,15 +18,26 @@ use Symfony\Component\DependencyInjection\Compiler\ServiceLocatorTagPass;
 
 class TranslatorPass implements CompilerPassInterface
 {
+    private $translatorServiceId;
+    private $loaderServiceId;
+    private $loaderTag;
+
+    public function __construct($translatorServiceId = 'translator.default', $loaderServiceId = 'translation.loader', $loaderTag = 'translation.loader')
+    {
+        $this->translatorServiceId = $translatorServiceId;
+        $this->loaderServiceId = $loaderServiceId;
+        $this->loaderTag = $loaderTag;
+    }
+
     public function process(ContainerBuilder $container)
     {
-        if (!$container->hasDefinition('translator.default')) {
+        if (!$container->hasDefinition($this->translatorServiceId)) {
             return;
         }
 
         $loaders = array();
         $loaderRefs = array();
-        foreach ($container->findTaggedServiceIds('translation.loader', true) as $id => $attributes) {
+        foreach ($container->findTaggedServiceIds($this->loaderTag, true) as $id => $attributes) {
             $loaderRefs[$id] = new Reference($id);
             $loaders[$id][] = $attributes[0]['alias'];
             if (isset($attributes[0]['legacy-alias'])) {
@@ -34,8 +45,8 @@ class TranslatorPass implements CompilerPassInterface
             }
         }
 
-        if ($container->hasDefinition('translation.loader')) {
-            $definition = $container->getDefinition('translation.loader');
+        if ($container->hasDefinition($this->loaderServiceId)) {
+            $definition = $container->getDefinition($this->loaderServiceId);
             foreach ($loaders as $id => $formats) {
                 foreach ($formats as $format) {
                     $definition->addMethodCall('addLoader', array($format, $loaderRefs[$id]));
@@ -44,7 +55,7 @@ class TranslatorPass implements CompilerPassInterface
         }
 
         $container
-            ->findDefinition('translator.default')
+            ->findDefinition($this->translatorServiceId)
             ->replaceArgument(0, ServiceLocatorTagPass::register($container, $loaderRefs))
             ->replaceArgument(3, $loaders)
         ;