+++ src/Symfony/Component/Debug/CHANGELOG.md
@@ -6,6 +6,11 @@ CHANGELOG
 
 * removed the symfony_debug extension
 
+3.4.0
+-----
+
+* deprecated `ErrorHandler::stackErrors()` and `ErrorHandler::unstackErrors()`
+
 3.3.0
 -----