+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -21,6 +21,12 @@ CHANGELOG
    has been dropped, use the `HttpKernel::allowCustomResponseCode()` method
    instead
 
+3.4.0
+-----
+
+ * added `AddCacheClearerPass`
+ * added `AddCacheWarmerPass`
+
 3.3.0
 -----