+++ src/Symfony/Component/Stopwatch/CHANGELOG.md
@@ -5,3 +5,5 @@ CHANGELOG
 -----
 
  * added the `Stopwatch::reset()` method
+ * allowed to measure sub-millisecond times by introducing an argument to the
+   constructor of `Stopwatch`